Refactor: python, cts: Move MaintenanceMode into the _cts module.
No code changes.
